fix(actions): rename to .svelte.ts + getter pattern for saving state
