+++ src/client/actions/registry.svelte.ts
